+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Released on ??
 - Implemented `From` trait for `Metadata`.
 - `create` and `append` will now return a `WriteStream` instead of a box, which will contain the inner stream which supports `Write` and may support `Seek` (according to the protocol).
 - `read` will now return a `ReadStream` instead of a box, which will contain the inner stream which supports `Read` and may support `Seek` (according to the protocol).
+- Metadata times (`created`, `accessed` and `modified`) are now `Option<SystemTime>` in order to provide the user to handle unset times, in case it is not supported by the remote server.
 
 ## 0.1.1